"move: domain 폴더로 이동"
