+++ src/test/java/domain/CarTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package domain;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -9,15 +9,16 @@ import racingcar.domain.Car;
 public class CarTest {
     @Test
     void 자동차_이름이_다섯_글자가_넘으면_예외를_던진다() {
-        Assertions.assertThrows(IllegalArgumentException.class, () -> new Car("abcdef"));
+        Assertions.assertThrows(IllegalArgumentException.class, () -> new Car("abcdef",0));
     }
 
     @Test
     void 주어진_숫가자_4이상이면_전진한다() {
-        Car car = new Car("박정언");
+        Car car1 = new Car("박정언",0);
+        Car car2 = new Car("홍길동",1);
 
-        car.moveOrStop(4);
+        car1.moveOrStop(4);
 
-        assertThat(car.getPosition()).isEqualTo(1);
+        assertThat(car1.compareTo(car2)).isEqualTo(0);
     }
 }
