+++ src/test/java/domain/RandomNumberGeneratorTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
